fix: responsive video component (#552)
Sorry for the large PR. It also contains the migration of youtube
embeds.

There are 2 files that are related to the new component: 
- `website/src/components/YoutubeEmbed.js`
- `website/src/theme/MDXComponents.js`

The rest of the changes comprise of migrating all current youtube
embeds. I used regex to find and replace them. I also checked through
all the changes manually to verify nothing went wrong and no content was
lost.

# Changes
- New responsive `<YoutubeEmbed />` component with lazy loading. (Lazy
loading improves initial time-to-interactive from >1.1s to 600-700ms).
- Component added to Global MDX Scope so that it doesn't have to be
imported in each markdown file.
- Migration of existing embeds to new component

- Made a generic wrapper component VideoEmbed that can be extended in the future to support more hosts.