feat: Table Widget Add Rows (#553)
Added docs for new feature: Adding new rows to Tables. Plan to rehaul
the inline-editing page, but just added these docs for now to meet the
release of the feature.

---

~~Currently a draft PR. Wanted to check in and see if this is an okay
way to add this information, or if I should take another route.~~

For now I've got the new feature information added to the inline-editing
page, because they share a lot of features/functionality and the work
very similarly. I've put it into its own subsection at the bottom,
because it has some unique properties that wouldn't make as much sense
mixed into the big properties table of the inline-editing section; there
is already quite a lot of information on the page and I think mixing the
two topics together would make it pretty hard to read.

To accomplish adding a new major section for the "Add new rows" feature,
I've bumped all heading levels on the page down by 1 (except for the
title). `##` to `###` and so-on. This actually makes the organization of
this page more consistent with the rest of the docs in terms of what
content is at what heading level.

That was more info than I intended to dump... But does this seem okay?

[First Draft,
notion](https://www.notion.so/appsmith/Table-add-new-row-c3ed0d4c5d1e4ee29077e084df3830b8)
[Current docs
(Table)](https://docs.appsmith.com/reference/widgets/table/)
[Current docs (Table >
inline-editing)](https://docs.appsmith.com/reference/widgets/table/inline-editing)